facade-service/requirements.txt
fastapi
uvicorn
pydantic
httpx
kafka-python
python-consul
